Added BotUtils class, refactored command file loading
