+++ src/BotClient.js
@@ -1,9 +1,8 @@
-const fs = require('node:fs');
-const path = require('node:path');
 const { Client, Collection, Events, GatewayIntentBits } = require('discord.js');
 const { token } = require('../auth.json');
 const cron = require("node-cron");
 const { CHANNELS } = require('./constants');
+const { BotUtils } = require('./Utils');
 
 class BotClient extends Client {
 
@@ -17,10 +16,9 @@ class BotClient extends Client {
     }
 
     init() {
-        this.commands = new Collection();
         this.once(Events.ClientReady, readyClient => {
             console.log(`Ready! Logged in as ${readyClient.user.tag}`);
-            this.scheduleMessage("element.title", "element.readabledate");
+            // this.scheduleMessage("element.title", "element.readabledate");
         });
         this.login(token);
         this.setCommands();
@@ -28,21 +26,10 @@ class BotClient extends Client {
     }
 
     setCommands() {
-        const foldersPath = path.join(__dirname, 'commands');
-        const commandFolders = fs.readdirSync(foldersPath);
-        for (const folder of commandFolders) {
-            const commandsPath = path.join(foldersPath, folder);
-            const commandFiles = fs.readdirSync(commandsPath).filter(file => file.endsWith('.js'));
-            for (const file of commandFiles) {
-                const filePath = path.join(commandsPath, file);
-                const command = require(filePath);
-                // Set a new item in the Collection with the key as the command name and the value as the exported module
-                if ('data' in command && 'execute' in command) {
-                    this.commands.set(command.data.name, command);
-                } else {
-                    console.log(`[WARNING] The command at ${filePath} is missing a required "data" or "execute" property.`);
-                }
-            }
+        this.commands = new Collection();
+        const commandList = BotUtils.getCommands();
+        for(let command of commandList) {
+            this.commands.set(command.data.name, command);
         }
     }
